Remove ban command and ban list command modules; refactor message controller to integrate admin commands.
